% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carima.R
\name{carima}
\alias{carima}
\title{Chatterjee Auto-Regressive Modeling}
\usage{
carima(data, silent = FALSE, no.predict = 0)
}
\arguments{
\item{data}{a vector of time series data.}

\item{silent}{whether or not to silence output.}

\item{no.predict}{number of predictions to make.}
}
\description{
This function models autoregressive processes as described by Dr. 
Arunendu Chatterjee (2016.12.09).
}
\examples{
carima()
}
\keyword{carima}

