% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csts.R
\name{csts}
\alias{csts}
\title{Seasonal Time Series Regression}
\usage{
csts(ts.data = NA, ts.start = NULL, ts.end = NULL, ts.frequency = 1,
  plot.initial = FALSE, out = c("data.frame", "model"), df.print = FALSE,
  no.predict = 0)
}
\arguments{
\item{ts.data}{time series data}

\item{ts.frequency}{the number of observations per unit of time}

\item{year.start}{the (optional) year in which the data starts}

\item{year.end}{the (optional) year in which the data ends}
}
\description{
This function performs time series regression with seasonal considerations by computing the
following (as per Dr. Chatterjee's instruction in Fall 2016): (centered) moving average,
seasonl ratio, raw seasonal indices, normalized seasonal indices, de-seasonalized raw data,
de-seasonalized predictions, and re-seasonalized predictions. Optionally, it can carry this
forecasting method out past the end of the data set.
. It either returns a data frame containing
an abundance of information (including the final prediction) or it returns the model.
}
\examples{
csts()
}
\keyword{csts}

