% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsr.R
\name{ctsr}
\alias{ctsr}
\title{Time Series Regression}
\usage{
ctsr(data, frequency = 1)
}
\arguments{
\item{data}{a vector of time series data.}

\item{frequency}{the number of observations per unit of time.}
}
\description{
This function performs time series regression with multiple methods, then it 
selects the best method based on mean absolute deviation. Once the best model
is selected, it returns some information about it.
}
\examples{
ctsr()
}
\keyword{ctsr}

