% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.poly.R
\name{which.poly}
\alias{which.poly}
\title{Polynomial Regression}
\usage{
which.poly(x, t = NULL, degree.max = 10)
}
\arguments{
\item{x}{a vector of time series data upon which to perform polynomial regression}

\item{t}{a vector of time series sequence information}

\item{degree.max}{the maximum degree to consider for polynomial regression}
}
\description{
This function performs time series polynomial regression. It returns the
polynomial regression with the best adjusted r-squared value.
}
\examples{
which.polt()
}
\keyword{which.poly}

