% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmlr.R
\name{cmlr}
\alias{cmlr}
\title{Chatterjee Multiple Linear Regression}
\usage{
cmlr(x)
}
\arguments{
\item{x}{a data frame for regression, where the first column is the dependent variable}
}
\description{
This function performs multiple linear regression (with transformations) using
the methodology taught by Dr. Arunendu Chatterjee for Fall 2016.
}
\examples{
cmlr()
}
\keyword{cmlr}

