% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csts.R
\name{csts}
\alias{csts}
\title{Seasonal Time Series Regression}
\usage{
csts(data = NA, start = NULL, end = NULL, frequency = 1,
  plot.initial = FALSE, out = c("data.frame", "model"),
  results.print = FALSE, no.predict = 0)
}
\arguments{
\item{data}{time series data.}

\item{start}{the time of the first observation. Either a single number or a vector of two
integers, which specify a natural time unit and a (1-based) number of samples into the time
unit. See the examples for the use of the second form.}

\item{end}{the time of the last observation, specified in the same way as start.}

\item{frequency}{the number of observations per unit of time.}

\item{plot.initial}{a boolean indicating whether you want a plot of the initial time series.}

\item{out}{whether you want a data frame (out = "data.frame") or the deseasonalized linear
model (out = "model") to be provided as output.}

\item{results.print}{a boolean value indicating whether you want the data frame to be printed.}

\item{no.predict}{the number of predictions to perform.}
}
\description{
This function performs time series regression with seasonal considerations by computing the
following (as per Dr. Chatterjee's instruction in Fall 2016): (centered) moving average,
seasonl ratio, raw seasonal indices, normalized seasonal indices, de-seasonalized raw data,
de-seasonalized predictions, and re-seasonalized predictions. Optionally, it can carry this
forecasting method out past the end of the data set.
. It either returns a data frame containing
an abundance of information (including the final prediction) or it returns the model.
}
\examples{
csts()
}
\keyword{csts}

