% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSmooth.R
\name{eSmooth}
\alias{eSmooth}
\title{Simplified Exponential Smoothing}
\usage{
eSmooth(x, alpha = 0.2)
}
\arguments{
\item{x}{a vector to be exponentially smoothed}

\item{alpha}{Holt-Winter smoothing alpha}
}
\description{
This function computes an exponentially smoothed time series with a 
given alpha by simplifying the process.
}
\examples{
eSmooth()
}
\keyword{eSmooth}

