% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.poly.R
\name{ts.poly}
\alias{ts.poly}
\title{Polynomial Time Series Regression}
\usage{
ts.poly(x, t = NULL, degree.max = 10)
}
\arguments{
\item{x}{a vector of time series data upon which to perform polynomial regression.}

\item{t}{a vector of time series sequence information.}

\item{degree.max}{the maximum degree to consider for polynomial regression.}
}
\description{
This function performs polynomial time series regression. It returns the
polynomial regression with the best adjusted r-squared value.
}
\examples{
ts.poly()
}
\keyword{ts.poly}

